+++ lib/appfuel/initializers/web_api.rb
@@ -1,4 +1,5 @@
 Appfuel::Initialize.define('global.web_api') do |config, container|
   require 'rest-client'
+  require 'appfuel/storage/web_api'
   container.register('web_api.http_adapter', RestClient)
 end