+++ lib/appfuel/storage.rb
@@ -1,5 +1,3 @@
 require_relative 'storage/repository'
 require_relative 'storage/file'
 require_relative 'storage/memory'
-require_relative 'storage/db'
-require_relative 'storage/web_api'