+++ lib/appfuel/storage/aws_dynamo_db.rb
@@ -1,0 +1,2 @@
+require_relative 'aws_dynamo_db/repository'
+require_relative 'aws_dynamo_db/no_sql_model'