+++ lib/appfuel/storage/aws_dynamo_db/repository.rb
@@ -1,0 +1,11 @@
+module Appfuel
+  module AwsDynamoDb
+    class Repository < Appfuel::Repository::Base
+      class << self
+        def container_class_type
+          "#{super}.aws.dynamo_db"
+        end
+      end
+    end
+  end
+end